fix(handler): ensure that code parameter is validated
This commit the following validation for the params related to
currencies code:

- Verify if they aren't an empty string.
- Verify if they follow the  ISO 4217, always having 3 digits.

+++ internal/handler/currency_handler.go
@@ -26,11 +26,14 @@ func (h *CurrencyHandler) ConvertCurrency(w http.ResponseWriter, r *http.Request
 	to := strings.ToUpper(r.URL.Query().Get("to"))
 	amountStr := r.URL.Query().Get("amount")
 
-	if from == "" || to == "" || amountStr == "" {
+	if from == "" || to == "" || (amountStr == "") {
 		commons.RespondWithError(w, http.StatusBadRequest, "Missing required parameters")
 		return
 	}
-
+	if len(from) != 3 || len(to) != 3 {
+		commons.RespondWithError(w, http.StatusBadRequest, "Invalid currency code, must be 3 characters long following ISO 4217")
+		return
+	}
 	amount, err := strconv.ParseFloat(amountStr, 64)
 	if err != nil {
 		commons.RespondWithError(w, http.StatusBadRequest, "Invalid amount")
@@ -61,7 +64,14 @@ func (h *CurrencyHandler) AddCurrency(w http.ResponseWriter, r *http.Request) {
 		commons.RespondWithError(w, http.StatusBadRequest, "Invalid request payload")
 		return
 	}
-
+	if currency.Code == "" || currency.Rate == 0 {
+		commons.RespondWithError(w, http.StatusBadRequest, "Invalid currency code or rate")
+		return
+	}
+	if len(currency.Code) != 3 {
+		commons.RespondWithError(w, http.StatusBadRequest, "Invalid currency code, must be 3 characters long following ISO 4217")
+		return
+	}
 	if err := h.currencyService.AddCurrency(r.Context(), currency.Code, currency.Rate); err != nil {
 		commons.RespondWithError(w, http.StatusInternalServerError, "Failed to add currency")
 		return
@@ -71,7 +81,16 @@ func (h *CurrencyHandler) AddCurrency(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *CurrencyHandler) RemoveCurrency(w http.ResponseWriter, r *http.Request) {
-	code := chi.URLParam(r, "code")
+	code := strings.ToUpper(chi.URLParam(r, "code"))
+
+	if code == "" {
+		commons.RespondWithError(w, http.StatusBadRequest, "Invalid currency code")
+		return
+	}
+	if len(code) != 3 {
+		commons.RespondWithError(w, http.StatusBadRequest, "Invalid currency code, must be 3 characters long following ISO 4217")
+		return
+	}
 
 	if err := h.currencyService.RemoveCurrency(r.Context(), code); err != nil {
 		commons.RespondWithError(w, http.StatusInternalServerError, "Failed to remove currency")
